+++ modules/Students/Student.php
@@ -79,7 +79,7 @@ else
 
 if ( User( 'PROFILE' ) !== 'admin' )
 {
-	$can_edit_from_where = " FROM PROFILE_EXCEPTIONS WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'";
+	$can_edit_from_where = " FROM profile_exceptions WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'";
 
 	if ( User( 'PROFILE' ) !== 'student'
 		&& ! User( 'PROFILE_ID' ) )
@@ -654,7 +654,7 @@ if (  ( UserStudentID()
 		|| User( 'PROFILE' ) === 'student' )
 	{
 		$can_use_sql = "SELECT MODNAME
-			FROM PROFILE_EXCEPTIONS
+			FROM profile_exceptions
 			WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'
 			AND CAN_USE='Y'
 			AND MODNAME LIKE 'Students/Student.php%'";